+++ plugins/http2_adapter/test/http2_test.dart
@@ -76,4 +76,45 @@ void main() {
     final res = await dio.post('post', data: 'TEST');
     expect(res.data.toString(), contains('TEST'));
   });
+
+  test('catch DioException when receiveTimeout', () {
+    final dio = Dio(
+      BaseOptions(
+        baseUrl: 'https://httpbun.com/',
+        receiveTimeout: Duration(seconds: 5),
+      ),
+    );
+    dio.httpClientAdapter = Http2Adapter(
+      ConnectionManager(
+        idleTimeout: Duration(milliseconds: 10),
+      ),
+    );
+
+    expectLater(
+      dio.get('/drip?delay=10&numbytes=1'),
+      allOf([
+        throwsA(isA<DioException>()),
+        throwsA(predicate(
+            (DioException e) => e.type == DioExceptionType.receiveTimeout)),
+        throwsA(predicate(
+            (DioException e) => e.message!.contains('0:00:05.000000'))),
+      ]),
+    );
+  });
+
+  test('no DioException when receiveTimeout > request duration', () async {
+    final dio = Dio(
+      BaseOptions(
+        baseUrl: 'https://httpbun.com/',
+        receiveTimeout: Duration(seconds: 5),
+      ),
+    );
+    dio.httpClientAdapter = Http2Adapter(
+      ConnectionManager(
+        idleTimeout: Duration(milliseconds: 10),
+      ),
+    );
+
+    await dio.get('/drip?delay=1&numbytes=1');
+  });
 }