+++ plugins/http2_adapter/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-*None.*
+- Implement `sendTimeout` and `receiveTimeout` for the adapter.
 
 ## 2.3.1+1